requirements.txt
# FinTech Forecasting Application - Dependencies
# CS4063 - Natural Language Processing Assignment 2

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Data Processing
numpy==1.26.4
pandas==2.2.2
yfinance==0.2.38

# Traditional Forecasting Models
statsmodels==0.14.1
scipy==1.12.0

# Neural Network Models (Deep Learning)
tensorflow==2.15.0
keras==2.15.0

# Scheduling
APScheduler==3.10.4

# Database
# SQLite is included with Python, no additional package needed
pymongo==4.6.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.9.0.post0

# Testing
pytest==7.4.3
unittest2==1.1.0

# Optional but recommended for production
gunicorn==21.2.0

